requirements_api.txt
absl-py==2.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
astroid==4.0.2
asttokens==3.0.1
astunparse==1.6.3
async-lru==2.0.5
attrs==25.4.0
babel==2.17.0
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==6.2.2
catboost==1.2.8
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.2
cramjam==2.11.0
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.0
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.124.0
fastjsonschema==2.21.2
fastparquet==2024.11.0
flatbuffers==25.9.23
fonttools==4.61.0
fqdn==1.5.1
fsspec==2025.10.0
gast==0.4.0
geopandas==1.0.1
google-api-core==2.8.2
google-auth==2.43.0
google-auth-oauthlib==0.4.6
google-cloud-bigquery==3.18.0
google-cloud-core==2.5.0
google-cloud-storage==2.14.0
google-crc32c==1.7.1
google-pasta==0.2.0
google-resumable-media==2.8.0
googleapis-common-protos==1.56.4
graphviz==0.21
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
ipdb==0.13.13
ipykernel==7.1.0
ipython==8.37.0
isoduration==20.11.0
isort==7.0.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.5.0
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.9
lark==1.3.1
libclang==18.1.1
lightgbm==4.6.0
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mccabe==0.7.0
mistune==3.1.4
narwhals==2.13.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
notebook_shim==0.2.4
numpy==1.23.5
nvidia-nccl-cu12==2.28.9
oauthlib==3.3.1
opt_einsum==3.4.0
overrides==7.7.0
packaging==25.0
pandas==1.5.3
pandocfilters==1.5.1
parso==0.8.5
path==17.1.1
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.0
plotly==6.5.0
pluggy==1.6.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
protobuf==3.19.6
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pygeohash==3.2.1
Pygments==2.19.2
pylint==4.0.4
pyogrio==0.12.1
pyparsing==3.2.5
pyproj==3.7.1
pytest==9.0.1
python-dateutil==2.9.0.post0
python-json-logger==4.0.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rpds-py==0.30.0
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
shapely==2.1.2
six==1.17.0
soupsieve==2.8
stack-data==0.6.3
starlette==0.50.0
# tensorboard==2.10.1
# tensorboard-data-server==0.6.1
# tensorboard-plugin-wit==1.8.1
# tensorflow==2.10.0
# tensorflow-estimator==2.10.0
# tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tomli==2.3.0
tomlkit==0.13.3
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
Werkzeug==3.1.4
wrapt==2.0.1
xgboost==1.7.6
gcsfs
